refactor: reorg GlobalGuiControl.lua to Player/
