+++ TTS_xwing/src/Player/GlobalGuiControl.lua
